Fix 32bit compilation problem.

+++ Quantize.h
@@ -316,7 +316,7 @@ std::pair<T,T> io_minmax(DataSrc<T> ds, Histogram<T, sz> histogram,
   while(iPos < iSize) {
     size_t n_records = ds.read((unsigned char*)(&(data.at(0))),
                                std::min((iSize - iPos)*sizeof(T),
-                                        iCurrentIncoreSize));
+                                        UINT64(iCurrentIncoreSize)));
     if(n_records == 0) {
       WARNING("Short file..");
       break; // bail out if the read gave us nothing.
